Refactoring
Refactor to use double types and add DTO validation

Refactored all power, efficiency, and cost properties and calculations to use double instead of decimal for consistency. Added data annotations to DTOs for input validation. Updated business logic to use double arithmetic and improved code readability with switch statements and code cleanup. Cleaned up using directives and exception handling. These changes enhance type consistency, validation, and maintainability.

+++ PowerPlantCodingChallenge/DTO/ProductionPlanResultItem.cs
@@ -8,6 +8,6 @@ namespace PowerPlantCodingChallenge.DTO
         public required string PowerPlantName { get; set; }
 
         [JsonPropertyName("p")]
-        public decimal Power { get; set; }
+        public double Power { get; set; }
     }
 }
